app_api/requirements.txt
aiogram
aiohttp
Django==3.1
django-cleanup
django-cors-headers
django-debug-toolbar
django-filter
django-json-widget
django-multiselectfield
django-rest-framework
djangorestframework
djangorestframework-simplejwt
djoser
Pillow
requests
psycopg2-binary
drf-yasg
django-split-settings
python-dotenv
django-ckeditor
